+++ Brocfile.js
@@ -1,14 +1,14 @@
 /* global require, module */
 
-var EmberAddon = require('ember-cli/lib/broccoli/ember-addon');
+var EmberApp = require('ember-cli/lib/broccoli/ember-addon');
 
-var app = new EmberAddon({
-  'ember-bootstrap-datetimepicker': {
-  	'importBootstrapCSS': true,
-    'importBootstrapJS': true
+var app = new EmberApp({
+  'ember-bootstrap-datetimepicker': { 
+        'importBootstrapCSS': true,
+    'importBootstrapJS': true,
+    'importFontAwesome': true
   }
 });
-
 
 // Use `app.import` to add additional libraries to the generated
 // output files.