+++ addon/components/bs-datetimepicker.js
@@ -1,15 +1,18 @@
 
 import Ember from 'ember';
+import DateTimePickerTextFieldMixin from 'ember-bootstrap-datetimepicker/mixins/datetimepicker_textfield';
 
 var computed = Ember.computed;
 var datetimepickerDefaultConfig = Ember.$.fn.datetimepicker.defaults;
 var isDatetimepickerConfigKeys = Ember.keys(datetimepickerDefaultConfig);
 
-
 var bsDateTimePickerComponent = Ember.Component.extend({
+  concatenatedProperties: ["textFieldClassNames"],
   classNames: ["bs-datetimepicker-component"],
-  textFieldClass: Ember.TextField,
+  textFieldClass: Ember.TextField.extend(DateTimePickerTextFieldMixin),
+  textFieldClassNames: ["form-control"],
   textFieldName: computed.alias("elementId"),
+  textFieldOptions:null,
   date: null,
   bsDateTimePicker: null,
 
@@ -18,6 +21,7 @@ var bsDateTimePickerComponent = Ember.Component.extend({
   maxDate: datetimepickerDefaultConfig["maxDate"],
   disabledDates:[],
   enabledDates:[],
+  dateIcon: "glyphicon glyphicon-calendar",
 
   disabled:false,
   open: false,
@@ -25,24 +29,27 @@ var bsDateTimePickerComponent = Ember.Component.extend({
 
 
   _initDatepicker: function() {
-
     var self = this;
     var bsDateTimePicker = this.$(".datetimepicker").datetimepicker(this._buildConfig());
     var bsDateTimePickerFn = bsDateTimePicker.data("DateTimePicker");
 
     this.set('bsDateTimePicker', bsDateTimePickerFn);
-    bsDateTimePickerFn.setDate(self.get("date"));
     
+    if(self.get("date") === undefined) {
+      bsDateTimePickerFn.date(null);
+    } else {
+      bsDateTimePickerFn.date(self.get("date"));
+    }
+
     this.$('input.floatlabel').floatlabel({
       slideInput: false,
       labelStartTop: '10px'
     });
-
     
     bsDateTimePicker.on("dp.change", function(ev) {
       if(Ember.isNone(ev.date)) {
         self.set("date", undefined);
-      } else {
+      } else if (!ev.date.isSame(self.get('date'))) {
         if(self.forceDateOutput) {
           self.set("date", ev.date.toDate());
         } else {
@@ -77,23 +84,38 @@ var bsDateTimePickerComponent = Ember.Component.extend({
   }.observes("open"),
 
   _minDateObserver: function() {
-    this.get("bsDateTimePicker").setMinDate(this.get('minDate'));
+    if(Ember.isNone(this.get('minDate'))) {
+      this.get("bsDateTimePicker").minDate(false);
+    } else {
+      this.get("bsDateTimePicker").minDate(this.get('minDate'));
+    }
   }.observes("minDate"),
 
   _maxDateObserver: function() {
-    this.get("bsDateTimePicker").setMaxDate(this.get('maxDate'));
+    if(Ember.isNone(this.get('maxDate'))) {
+      this.get("bsDateTimePicker").maxDate(false);
+    } else {
+      this.get("bsDateTimePicker").maxDate(this.get('maxDate'));
+    }
   }.observes("maxDate"),
 
   _disabledDatesObserver: function() {
-    this.get("bsDateTimePicker").setDisabledDates(this.get('disabledDates'));
+    this.get("bsDateTimePicker").disabledDates(this.get('disabledDates'));
   }.observes("disabledDates"),
 
   _enabledDatesObserver: function() {
-    this.get("bsDateTimePicker").setEnabledDates(this.get('enabledDates'));
+    this.get("bsDateTimePicker").enabledDates(this.get('enabledDates'));
   }.observes("enabledDates"),
 
   _dateObserver: function() {
-    this.get("bsDateTimePicker").setDate(this.get('date'));
+    var bsDateTimePickerFn = this.get("bsDateTimePicker");
+
+    if(this.get("date") === undefined) {
+      bsDateTimePickerFn.date(null);
+    } else {
+      bsDateTimePickerFn.date(this.get("date"));
+    }
+
   }.observes("date"),
 
   _destroyDatepicker: function() {
@@ -112,11 +134,58 @@ var bsDateTimePickerComponent = Ember.Component.extend({
     toggleOpen: function() {
       this.toggleProperty('open');
     }
+  },
+
+  /**
+
+    Exposing the textField properties.
+    Every property beginning with "textField" will be exposed to the TextField view.
+
+    ```handlebars
+    {{bs-datetimepicker textFieldPlaceholder="Date"}}
+    ```
+    "textFieldPlaceholder" will be exposed to the TextField as "placeholder" property.
+
+  */
+  setUnknownProperty: function(key, value) {
+    var prop;
+    var ckey;
+    if(key.indexOf("textField") === 0) {
+
+      if(Ember.isNone(this.get('textFieldOptions'))) {
+        this.set('textFieldOptions',{});
+      }
+
+      if(Ember.IS_BINDING.test(key)) {
+        prop = key.substring(0,key.length-7);
+      }
+      else {
+       prop = key.substring(0,key.length);
+      }
+
+      ckey = prop.substring(9);
+      ckey = ckey.charAt(0).toLowerCase() + ckey.substr(1);
+
+      if(Ember.isNone(this.get('textFieldOptions.'+prop))) {
+
+        this.set('textFieldOptions.'+prop,ckey);
+
+        Ember.defineProperty(this, prop, null, value);
+      }
+    }
+    else {
+      if(Ember.platform.hasPropertyAccessors) {
+        Ember.defineProperty(this, key, null, value);
+      }
+      else {
+        this[key] = value;
+      }
+    }
   }
 });
 
 
-var computedProps = ["minDate","maxDate","disabledDates","enabledDates"];
+var computedProps = ["minDate","maxDate","disabledDates","enabledDates","dateIcon"];
 var newClassConfig = {};
 for(var i=0; i<isDatetimepickerConfigKeys.length; i++) {
   if(!computedProps.contains(isDatetimepickerConfigKeys[i])) {
@@ -128,4 +197,3 @@ bsDateTimePickerComponent.reopen(newClassConfig);
 
 
 export default bsDateTimePickerComponent;
-