+++ index.js
@@ -1,3 +1,4 @@
+/* jshint node: true */
 'use strict';
 
 var path = require('path');
@@ -28,11 +29,22 @@ module.exports = {
     	app.import(path.join(bootstrapPath, 'fonts/glyphicons-halflings-regular.svg'), { destDir: '/fonts' });
     	app.import(path.join(bootstrapPath, 'fonts/glyphicons-halflings-regular.ttf'), { destDir: '/fonts' });
     	app.import(path.join(bootstrapPath, 'fonts/glyphicons-halflings-regular.woff'), { destDir: '/fonts' });
+      app.import(path.join(bootstrapPath, 'fonts/glyphicons-halflings-regular.woff2'), { destDir: '/fonts'});
     }
 
 
 		//Import css from bootstrap-datetimepicker
 		app.import(path.join(app.bowerDirectory, '/eonasdan-bootstrap-datetimepicker/build/css/bootstrap-datetimepicker.css'));
+
+
+  if(options.importFontAwesome) {
+    app.import(app.bowerDirectory + '/font-awesome/css/font-awesome.min.css', { destDir: '/fonts'});
+    app.import(app.bowerDirectory + '/font-awesome/fonts/fontawesome-webfont.woff', { destDir: '/fonts'});
+    app.import(app.bowerDirectory + '/font-awesome/fonts/fontawesome-webfont.woff2', { destDir: '/fonts'});
+    app.import(app.bowerDirectory + '/font-awesome/fonts/fontawesome-webfont.eot', { destDir: '/fonts'});
+    app.import(app.bowerDirectory + '/font-awesome/fonts/fontawesome-webfont.svg', { destDir: '/fonts'});
+    app.import(app.bowerDirectory + '/font-awesome/fonts/fontawesome-webfont.ttf', { destDir: '/fonts'});
+  }
 
 
     //Import js from moment