+++ CHANGELOG.md
@@ -6,3 +6,23 @@
 ## 0.2.7
 
 * Introduced new `forceDateOutput` option
+
+## 0.2.8
+
+* update to eonasdan-bootstrap-datetimepicker v4.0.0
+
+
+## 0.3.0
+
+* Update to ember-cli v0.1.7
+* Prevent update of date if it didn't changed
+* Add the ability to override the form-group-addon's child span class
+
+## 0.3.1
+
+* Possibility to overwrite TextField with a template block
+* Exposing textField properties
+
+## 0.3.2
+
+* Upgrade to ember-cli 0.2.0